Refactor media data fetching to custom hook
